Send some important headers in a fixed order, sort the rest.
Go maps intentionally do not have consistency in the order of iteration over
keys and values. To get that, we need to create our own order of iteration.
A nice start is several key headers: Subject, From, Date, Message-Id and
References. These headers are output before other headers in a fixed order,
while the remaining is output sorted by keys.